CN-4783: sending Appsflyer CUID (#595)
